refactor(tests): split DomainProvider into focused Faker providers
Replace the monolithic DomainProvider with three focused providers that
each own a single API area:

- GoogleAuthProvider  – googleCredentials() fixture
- PurchaseProvider    – subscription/product tokens and purchase payloads
- RtdnProvider        – cloud messages and RTDN notification payloads

MonetizationProvider is unchanged. Faker::create() registers all four
providers; the @mixin docblock is updated accordingly.

A new FakerTest asserts that every expected provider is registered, the
old DomainProvider mixin is removed, and each provider's representative
method returns the expected type.

Closes #301
